changes tutorial1 CommentBox component to use JSX Syntax
From the tutorial:

```
The first thing you'll notice is the XML-ish syntax in your JavaScript.
We have a simple precompiler that translates the syntactic sugar to this
plain JavaScript[...]

Its use is optional but we've found JSX syntax easier to use than plain
JavaScript.
```

I'm not convinced, and they may not be, either— it looks like the
remainder of the tutorial uses the XML-style tags.

So, expect this commit to be immediately reverted, but trying as
an exercise, and commiting so one can look at the diffs.